exynos: Add concept of a "bad wake" GPIO to machine parameters
An invalid wakeup happens when the state of a certain GPIO is high
during wakeup.  This is used to detect a malicious user who might be
trying to mess with the "is sleeping" GPIO.  We detect this in U-Boot
upon wakeup and reboot if we find this attack.

BRANCH=none
BUG=chrome-os-partner:20172
TEST=Run the platform_TpmSleepTestServer autotest on pit (with all CLs
together)

[dianders: this is based on an old change to chromeos-v2011.12; see
Original-Change-Id.]

Change-Id: If00950313d3009b913b47fb77dbc27ef5c905112
Original-Change-Id: If97ccd23ea71619697e9c8d37e5b4967fec77a95
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58770
Reviewed-by: Simon Glass <sjg@chromium.org>