exynos: Add "bad wake" GPIO detection into wakeup path
An invalid wakeup happens when the state of a certain GPIO is high
during wakeup.  This is used to detect a malicious user who might be
trying to mess with the "is sleeping" GPIO.  We detect this in U-Boot
upon wakeup and reboot if we find this attack.

BRANCH=none
BUG=chrome-os-partner:20172
TEST=Run the platform_TpmSleepTestServer autotest on pit (with all CLs
together)

[dianders: this is based on an old change to chromeos-v2011.12, though
reworked significantly]

Change-Id: Id72908e6a435a07e637664e58a97d54c3191e516
Original-Change-Id: Icd020a956f118f488994731cc30704a77c405bd4
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58771
Reviewed-by: Simon Glass <sjg@chromium.org>