Add verified boot information and test
Add a description of how to implement verified boot using signed FIT images,
and a simple test which verifies operation on sandbox.

The test signs a FIT image and verifies it, then signs a FIT configuration
and verifies it. Then it corrupts the signature to check that this is
detected.

Series-to: u-boot
Series-version: 2
Series-process-log: sort, uniq
Series-changes: 2
- Use U-Boot's -c option instead of hard-coding a boot script
- Update README to fix typos

Series-cc: tomr, wolfgang, vadim
Series-cc: Kees Cook <keescook@chromium.org>
Series-cc: Will Drewry <wad@chromium.org>
Series-cc: Randall Spangler <rspangler@chromium.org>
Series-cc: Bill Richardson <wfrichar@chromium.org>
Series-cc: Andreas Bäck <andreas.back778@gmail.com>
Series-cc: Joel A Fernandes <joelagnel@ti.com>
Series-cc: u-boot-review@google.com

Cover-letter:
Verified boot implementation based on FIT
This series implemented a verified boot system based around FIT images
as discussed on the U-Boot mailing list, including on this thread:

http://permalink.gmane.org/gmane.comp.boot-loaders.u-boot/147830

RSA is used to implement the encryption. Images are signed by mkimage
using private keys created by the user. Public keys are written into
U-Boot control FDT (CONFIG_OF_CONTROL) for access by bootm etc. at
run-time. The control FDT must be stored in a secure place where it
cannot be changed after manufacture. Some notes are provided in the
documentaion on how this can be achieved. The implementation is fairly
efficient and fits nicely into U-Boot. FIT plus RSA adds around 18KB
to SPL size which is manageable on modern SoCs.

When images are loaded, they are verified with the public keys.

Some minor restructuring of the image code is included in this series,
since we now support signatures as well as hashes.

It is important to have a test framework for this series. For this, sandbox
is used, and a script is provided which signs images and gets sandbox to
load them using a script, to check that all is well. So some of the patches
here related to adding image support for sandbox. A follow-on series is
somewhat more agressive in further refactoring the FIT image support to
clean it up improve maintainability.

Rollback prevention has been added in a separate TPM patch. This ensures
that an attacker cannot boot your system with an old image that has been
compromised. Support for this is not built into bootm, but instead must
be scripted in U-Boot. It is possible that a standard scheme for this could
be devised by adding version number tags to the signing procedure. However
scripts do provide more flexibility. The TPM patch is here:

http://patchwork.ozlabs.org/patch/224163/

This series is available at:

http://git.denx.de/u-boot-x86.git

in the branch 'vboot'.

I have received a number of off-list comments - please do copy the list when
replying so that everyone can see your comments.
END

Change-Id: I436b5405d9dab95b1414ff0d58d1219074451edb
Signed-off-by: Simon Glass <sjg@chromium.org>