x86: Support loading kernel setup from a FIT
Add a new setup@ section to the FIT which can be used to provide a setup
binary for booting Linux on x86. This makes it possible to boot x86 from
a FIT.

Series-to: u-boot
Series-cc: vadim
Series-prefix: RFC
Cover-letter:
RFC: FIT images and device tree on x86
This little series adds support for booting FITs on x86. A FIT is a single
file that can contain a kernel, device tree and ramdisk.

This is a bit of an experiment, and is work in progress. I would like to
get feedback on the approach.

x86 kernels require a setup.bin file to boot. Rather than try to pack this
into the image in a clever way, this series just adds it as another image
in the FIT. At present compression is not supported for the setup binary.

The kernel image can be compressed using FIT's normal compression options,
such as bzip2 and lzo. U-Boot will decompress it as part of the 'bootm'
command.

More work is needed to make this implementation also support a zImage
(which contains a setup.bin and a compressed kernel in a single image), and
to pass the correct device tree binary to the kernel.

This series builds on the image and FIT code improvements that were
done for verified boot. The full series is available at:

http://git.denx.de/u-boot-x86.git

in branch 'setup'.
END

Change-Id: I1614bf7e588c3e1236502ca7773ab820843f2bd6
Signed-off-by: Simon Glass <sjg@chromium.org>