fdt: Makefile: Add CPP preprocessing to DTS Makefile
Similar to what's done upstream, dtc is preprocessed/filtered to
allow the #include usage, and to remove the DTC: dts>dtb message
that causes Makeall to say that the build failed.

Change-Id: I3675f3068609d7e507df79aa99331e5028a10a22
Signed-off-by: Tom Warren <twarren@nvidia.com>

Merged in some additional changes so that we can deal with upstream device
tree files which require CPP, but still produce a sensible error message
(i.e. file and line number) when the compilation files.

Signed-off-by: Simon Glass <sjg@chromium.org>