Move AP init code to arch/arm/cpu
This patch moves the AP init code to arch/arm/cpu.

Lowlevel_init.S assembly code is converted to .c version Also based on
OM pin configuration in UM, values for the corresponding boot modes
are set, e.g for emmc pin setting is 001000 -> OM[6:1] which comes out
to be 8.

vbendeb: note that console support in SPL is being dropped from the
original patchset. It was not working in the first place and will have
to be added later.

BUG=none
TEST=Compile ok.U-Boot booting fine on SMDK5250

Change-Id: I2478c6cd0b990d589994a7632fe192eb91e47e6f
Signed-off-by: Rajeshwari Shinde <rajeshwari.s@samsung.com>
Reviewed-on: https://gerrit-int.chromium.org/33647
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Simon Glass <sjg@google.com>

Modified to build on new upstream:
Signed-off-by: Simon Glass <sjg@chromium.org>