EXYNOS: SPI: Support SPI_PREAMBLE mode
Support interfaces with a preamble before each received message.

We handle this when the client has requested a SPI_XFER_END, meaning
that we must close of the transaction. In this case we read until we
see the preamble (or a timeout occurs), skipping all data before and
including the preamble. The client will receive only data bytes after
the preamble.

Series-to: u-boot
Series-cc: goog
Series-version: 2
Series-changes: 2
- Remove preamable_count variable which is not really needed
- Fix checkpatch warning (multiple assignments)

Change-Id: If6673d4bbf52c55484cb67606ad4c37c97276d5e
Signed-off-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Rajeshwari Shinde <rajeshwari.s@samsung.com>