exynos: Make sure ps_hold gets set in the SPL
Setting ps_hold ought to be one of the first things we do when we
first boot up.  If we wait until the main u-boot runs we won't set it
in time and the PMIC may power us back off.

I've moved the ps_hold setup into the generic power_init() which
should contain code that's currently duplicated in the
board_power_init() of several boards.

BUG=None
TEST=Can now boot from SPI flash

Change-Id: I5d41daabeed67f82e11729db510dc492e444942f
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/34416
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>