temp: exynos: Implement U-Boot size parameters in machine_params
At present this parameter is ignored in spl_load, and the value is
hard-coded to 512KB. Implement this feature so that we can support
larger U-Boot images (perhaps containing a payload to be written
to SPI flash), and smaller U-Boot images (i.e. not reading past the
end of the U-Boot image when it is smaller than 512KB.

Also make use of the MMC_MAX_BLOCK_LEN constant.

BUG=chrome-os-partner:18451
TEST=See that flasher no longer gives crc32 errors when run

Use this to write to SD card:

$ cros_bundle_firmware -b peach_pit -s -M exynos -F spi -w sd:.

Then insert SD card in pit, boot from SD and see that it flashes correctly.

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: Id6b039fd98a51e2088bacb226ea42a84092e1aa1
Reviewed-on: https://gerrit-int.chromium.org/34642
Reviewed-by: Vadim Bendebury <vbendeb@google.com>
Commit-Queue: Simon Glass <sjg@google.com>
Tested-by: Simon Glass <sjg@google.com>