Use proper device tree on peach
Make the peach device tree include all necessary components in
preparation to verified boot. Note that the 5420 SPL size is is 16K,
not 32K as the 5250 one. Let's override it for now in the peach
specific device tree.

The firmware-storage alias is being removed as it is not used anywhere
in the code (least yet) and the node it aliases does not exist either.

BUG=chrome-os-partner:18329
TEST=manual
  . build peach firmware image and program it on an SD card:

  $ emerge-peach_pit chromeos-bootimage peach-u-boot
  $ cros_write_firmware -b peach_pit -w sd:. -M exynos \
   -i /build/peach_pit/firmware/nv_image-peach_pit.bin -D \
  --dt /build/peach_pit/firmware/dts/exynos5420-peach_pit.dts

  . try booting a peach device from the SD card, observe it boot.

Change-Id: I42e6e04a1091025bca370508d2de48051cfa5642
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/35024
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Simon Glass <sjg@google.com>
Reviewed-by: Doug Anderson <dianders@google.com>