Provide facility to control s2mps11 clocks
The s2mps11 PMIC device is capable of driving three independent 32KHz
clocks. One of the clocks is used by the Infineon TPM in the peach-pit
design, the clock needs to be enabled at startup.

This change adds a property, named 'u-boot-clocks' to the s2mps11
device tree node. This property is a bitmap of the three clocks.

During initialization on peach-pit clock zero is set to follow the
default, and clock 1 is set for the Infineon TPM.

To facilitate access to the device tree node of the PMIC, the node
offset is being added to struct pmic.

BUG=chrome-os-partner:18329
TEST=manual
   . build and boot peach-pit u-boot. observe no s2mps11
     initialization problems reported. A later change adds support for
     the Infineon TPM which uses the clocks enabled by this change.

Change-Id: I723e79e4568e4abebeee545916dab153fbe81932
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/35239
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>