Do not flood console if LCD is not connected
In case LCD is not connected, the code persists trying to read the
EDID, flooding the console with error messages along the way.

There is no point in continuing if LCD is not available, this change
limits the amount of messages showing up on the console to one.

BUG=chrome-os-partner:18329
TEST=manual

   . built the new image, observed it still comingh up with LCD
     operational on snow. When bringing up with LCD disconnected, a
     single message shows up on the console:

      DP exynos_dp_read_edid() failed

Change-Id: I21df0c2d5507899d565d81f70b02f3fb9b46b9dd
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/35210
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Simon Glass <sjg@google.com>
Tested-by: Naveen Krishna Chatradhi <naveen@chromium.org>
Commit-Queue: David James <davidjames@google.com>