SMDK5420: Add LCD support for eDP panel
This patch adds LCD and backlight poweron code for eDP panel
on SMDK5420. Since the LCD and backlight settings are different
for SMDK5420 and peach, we are forced to define seperate
exynos_fb callbacks.
Using CONFIG_CHROMEOS_PEACH as of now to define seperate
callbacks for the two boards.

BUG=None
TEST=manual:
  before this fix
    "make smdk5420_config; make;" Gives the error:
    "smdk5420.c: In function exynos_lcd_power_on: smdk5420.c:263:25:
    error: ANX1120_I2C_BUS undeclared (first use in this function)"
  after this fix
   "make smdk5420_config; make;" No compilation errors.

Change-Id: I6a951033430de9c568ca5397f81fd5f9afcae120
Signed-off-by: Ajay Kumar <ajaykumar.rs@samsung.com>
Reviewed-on: https://gerrit-int.chromium.org/35252
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>