temp: Modify SPI layout to allow for larger u-boot images
When TPM and VBOOT code are compiled in, the u-boot.bin blob size
(which happens to include two EC binaries, both ro and rw) crosses the
room allocated for it in the memory layout.

Eventually we will need to trim the image size, but to keep
development going let's increase the u-boot blob space temporarily.

BUG=chrome-os-partner:18329, chrome-os-partner:18794
TEST=manual
   . built and booted to console u-boot on daisy, smdk5420 and peach

Change-Id: Id5712a0eb78396d02952525d253e1c12bca3fad3
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/35880
Reviewed-by: Simon Glass <sjg@google.com>
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>