temp: HACK off A7s early in the SPL
It is believed that the A7s are coming alive somehow and interfering
with boot.  Normally this causes problems in the kernel, but it's
conceivable that it could even cause problems in U-Boot.  Hack the
power off for each A7 core early in bootup.  With this change we no
longer get random crashes in the kernel boot.

This change is a temporary HACK because at some point I'd imagine that
we'll be doing more init of the various cores in U-Boot.  When we do
that we can include this initialization there.  We also probably need
a new definition for the exynos5_power structure (since it's different
for 5420).

BUG=chrome-os-partner:18508
TEST=Boot lots of times with 'maxcpus=4 slub_debug=FZPUA' and don't
see console gibberish around the time that cores come up.  Don't see
lots of random crashes.

Change-Id: I6042af8f28491a4ae1c46a1250fdc0ad9bcda405
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/36000
Tested-by: Doug Anderson <dianders@google.com>
Reviewed-by: Olof Johansson <olofj@google.com>
Reviewed-by: Sonny Rao <sonnyrao@google.com>
Reviewed-by: Simon Glass <sjg@google.com>
Reviewed-by: Vadim Bendebury <vbendeb@google.com>
Commit-Queue: Doug Anderson <dianders@google.com>