exynos: Avoid function instrumentation for microsecond timer
For tracing to work it has to be able to access the microsecond timer
without causing a recursive call to the function entry/exit handlers.
Add attributes to the relevant functions to support this.

Series-to: u-boot
Series-cc: tomr, u-boot-review@google.com
Series-prefix: RFC
Cover-letter:
RFC: Add tracing functionality to U-Boot
This series adds a tracing feature to U-Boot which is useful for
profiling boot time and other purposes.

The core trace library relies on standard gcc function instrumentation
and a microsecond timer which should work correctly on almost any
architecture. Generic board is used to avoid the need to add the same
code in multiple places (CONFIG_SYS_GENERIC_BOARD).

Tracing must be built into the U-Boot image at build time, but can be
paused and resumed while running. A trace buffer is used to collect
trace information. This buffer can then be transmitted to a host for
processing. A host-based processing tool is provided which converts the
data to the same format used by Linux, and that can be read by the
pytimechart GUI tool.

A U-Boot 'trace' command provides access to the trace information,
including support for writing it to memory in a few forms.

Support is provided for stopping tracing at the last possible moment in
the bootm process (just before U-Boot jumps to the OS). This is done with
a new 'fake go' command, which allows bootm to go through the motions of
an OS boot without actually committing to it. Once the 'fake go' is
complete, U-Boot can stop tracing and continue execution to transmit the
trace information to the host, before jumping to the OS for real.

The system is tested on sandbox and has previously worked on ARM Tegra
and Exynos machines. A simple test script is provided for sandbox, along
with documentation on how to use the feature.

This feature has been successfully used alongside bootstage to locate
bottlenecks and reduce overall boot time.

Because of the bootm changes this series sits on top of the image
improvements that were done for verified boot. This series is available
at:

   http://git.denx.de/u-boot-x86.git

in branch 'trace'.
END
Change-Id: Ic5846bb2adc0d861f0f3f43b9bda513053fc7284
Signed-off-by: Simon Glass <sjg@chromium.org>