FIX: video: exynos_fb: Remove callbacks from the driver
Replaced the functionality of callbacks by using a standard set of functions.
Instead of implementing and hooking up a callback, put the same code in one of
the standard set of functions by overriding it.

This patch is tested only on SMDK5250.
For Trats and universal_c210 board, it is only compile tested.

FIX: This patch has a bug: exynos_backlight_on() weak attribute is wrong.

Signed-off-by: Ajay Kumar <ajaykumar.rs@samsung.com>
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=none
TEST=none
BRANCH=peach

Change-Id: I1c81d21e2294c484a4c6b8c355ddbec3ee03e6b5
Reviewed-on: https://gerrit-int.chromium.org/36091
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>