Rev 1 peach device tree
This change modifies the device tree to reflect the rev 1
hardware.

Eventually the EC will be accessible through SPI interface, but for
early bringup we might want to use the i2c connection (as reflected in
this version of the device tree). The PMU is left out for now (as it
is connected to the EC).

BUG=chrome-os-partner:18329
TEST=none

Change-Id: I93cfc8aebaf59be2b0a8d9f13518364d2684944c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/36618
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Doug Anderson <dianders@google.com>
Add a pointer to the correct serial console so that we get serial output.
Signed-off-by: Simon Glass <sjg@chromium.org>