Specify memory type in the device tree
The SPL must know the type of DDR3 it is supposed to initialize. The
type is communicated through the machine_param structure prepared by
the cros_bundle_firmware utility based on the device tree contents.

Set the default type to ddr3 and override it to lpddr3 for peach-adv
and smdk5420 platforms.

BUG=chrome-os-partner:18329
TEST=manual
   . built images for smdk5420 and adv with and without patches
     introducing DDR3 support (in addition to LPDDR3). All systems
     booted as expected.

Change-Id: I0f3285969a3993b42f66e93e0c0f025dcc4c7cc1
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/36669
Reviewed-by: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>
Commit-Queue: Vadim Bendebury <vbendeb@google.com>