exynos: Add a common DT based PMIC driver initialization
Most of i2c PMIC drivers follow the same initialization sequence,
let's generalize it in a common file.

The initialization function finds the PMIC in the device tree, and if
found - registers it in the list of known PMICs and initializes it,
iterating through the table of settings supplied by the caller.

For now only s2mps11 is being moved to the new scheme. To simplify
matters, the 32KHz clock output required by the TPM is enabled through
an initialization table entry, which obviates the need for a separate
routine to enables the clock as well as a device-tree property.

BUG=chrome-os-partner:18329
TEST=boots fine on peach-adv, the TPM still works.

Change-Id: I3454677f93675288d026554ecf9b520979853b6a
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/36810
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Doug Anderson <dianders@google.com>
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Modified to include pmic_common only when CONFIG_POWER is defined (build
breakage)
Signed-off-by: Simon Glass <sjg@chromium.org>