exynos: Use common pmic_reg_update() definition
This function is used by different Exynos platforms, put it in the
common file. In fact the function name us misleading, as it can not
updates bits from one to zero, only from zero to one. This will be
addressed under a separate change.

BUG=chrome-os-partner:18329
TEST=manual
   . smdk5250 builds and boots fine

Change-Id: I0ecef3db56b28dec17849710ffb136df7e7d45e3
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/36696
Reviewed-by: Simon Glass <sjg@google.com>
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Doug Anderson <dianders@google.com>