arm: Enable -ffunction-sections / -fdata-sections / --gc-sections
While other architectures have enabled these gcc / ld options for some
time on U-Boot itself, ARM has only been doing this on SPL.  Enable this
on full U-Boot as well now.

This is a pending patch from:

http://patchwork.ozlabs.org/patch/239574/

It saves around 40KB of text in U-Boot and 2KB of data on peach_pit.

BUG=none
BRANCH=none
TEST=manual:
Build and boot on pit and see that things seem to work as before.

$ crosfw -Vzw

Original size:
   text	   data	    bss	    dec	    hex	filename
 443082	  19040	 303648	 765770	  baf4a	b/chromeos_peach/u-boot
   5808	   1212	      0	   7020	   1b6c	b/chromeos_peach/spl/u-boot-spl

New sizes:
   text	   data	    bss	    dec	    hex	filename
 404473	  17560	 303308	 725341	  b115d	b/chromeos_peach/u-boot
   5764	   1212	      0	   6976	   1b40	b/chromeos_peach/spl/u-boot-spl

Change-Id: I9d850303d8ba56c8ef60f47023a187e5f6963762
Cc: Albert ARIBAUD <albert.u.boot@aribaud.net>
Signed-off-by: Tom Rini <trini@ti.com>
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57639
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>