cros: Use UUID for root partition
This makes it so that the enumeration order of potential boot devices in
u-boot is independent of the enumeration order of these same devices in
the kernel.

Testing was done with the puppy build.

BUG=none
TEST=Was able to boot with SD card, internal MMC, and USB thumb drive.
Confirmed that SD card and USB took priority over internal MMC, as expected.
Observed kernel command line during boot and saw that UUID was used for root.

Change-Id: Idc1c4cb92d705738928670449d03aa515ab3e3d9
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/49518
Reviewed-by: Jimmy Zhang <jimmzhang@nvidia.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Tom Warren <twarren@nvidia.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/60590
Commit-Queue: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>