elog: Add event log code as written for coreboot
This code maintains and extends an event log in a section of flash.
The format of the log is somewhat following the SMBIOS standard
(http://www.dmtf.org/standards/smbios) with certain deviations.

A fairly close implementation specification could be found in
https://code.google.com/p/firmware-event-log/wiki/FirmwareEventLogDesign.

The main discrepancies are as follows:

 - double buffer is not implemented, i.e. when the log grows close to
   capacity, it is shrunk and rewritten over the same aread in flash.

 - as a result, the log header structure does not include the
   elog_sequence field.

 - the space of log events is extended to include the FW Wake event
   (the document above does not inlude many recent event type
   additions).

This submission is a copy of working firmware elog implementation from
coreboot, it is being submitted as is (and can not be compiled/run
yet) for easier tracking of changes.

BRANCH=pit
BUG=chrome-os-partner:16132
TEST=With this and other changes, built and booted on pit. Ran mosys
  eventlog list and saw that the expected events with the expected
  timestamps were in the eventlog BRANCH=pit

Change-Id: I5b97ea6c1bf7cbc24a3064a05cf60263dcf296d1
Signed-off-by: Gabe Black <gabeblack@google.com>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/66771