gen: add 'describe' option to the gpio command
GPIO configuration on certain platforms is quite involved. It is
convenient to be able to query a certain GPIO pin properties at run
time.

This change adds the ability to invoke a platform provided function,
if it is defined. This function is expected to print on the console
any information relevant to the pin passed in as the parameter.

BUG=chrome-os-partner:18329
TEST=manual
   . tested when the exynos specific implementation was added

Change-Id: I9de837806a4bd975a46cf21ecef39ef9ccbc1eab
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/37102
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Simon Glass <sjg@google.com>