gen: provide support for SPI half duplex slaves
SPI bus provides for full duplex communications, but some slave
devices are not designed to support or require it: they wait for the
command first and then they respond to the command at some point in
the future.

Existing u-boot SPI drivers do not work well in this case: they can
send and transmit data independently, but they do not have means of
determining when the slave is actually ready to respond to a command.

Luckily, slave devices which operate asynchronously often encapsulate
their responses sent to the master in frames, which start with a well
defined byte and include a checksum.

One way of supporting these slaves is to continue clocking the bus
once master completed transmission of the command. The master receiver
should enter the so called hunt mode at this point, where it keeps
sampling data from the idle bus, until the slave frame start byte is
seen. This would synchronize master to the slave frame, after which
the receiver would receive the remaining number of frame bytes and
pass it up the stack for verification/processing.

Two additional variables are required for this operation: the value of
the frame start byte and the maximum timeout the receiver should wait
for the response.

This change adds the necessary fields and an interface declaration to
support SPI half duplex mode.

BUG=chrome-os-partner:18329
TEST=manual
  . no functional testing yet, just made sure that peach and snow
    u-boot still build and run

Change-Id: I15fb0bec7a08a8d9c3109ec8fc3199871741a0e8
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/37573
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>