elog: Make it possible to build the event log into SPL
Only the SPL part of firmware participates in resume, so it's where we
need to record a resume event. This change makes it possible to build
the event log into the SPL.

Also, in the SPL, we can't copy the contents of flash into memory like the
elog code normally would. Instead, we have to parse the log manually. Some of
the event log code is still useful, though, specifically the code that puts
together an event. Only that code is included when building the SPL.

To make it easier to support the event log with custom code in the spl, some
data structures are moved from the internal header file to the external
header.

BUG=chrome-os-partner:16132
TEST=Built and booted on pit.
BRANCH=pit

Change-Id: I824c451263cc4b8d259fcc5cf069a1674a1be2d0
Signed-off-by: Gabe Black <gabeblack@google.com>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/66776