cros: Clean up SPI EC driver layering
The actual overhead on the SPI physical layer is four bytes: framing,
status and length in the beginning and checksum in the end. The SPI
driver consumes the framing byte and passes the rest to the EC driver
layer for verification and stripping of the rest of the framing.

This also requires certain alignment adjustments, even though not
essential in case of EC.

BUG=chrome-os-partner:18329
TEST=manual
   . built and booted peach - the EC communications work fine.

Change-Id: I0e22523e5418562484ef597fb44a2f986e649e68
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/37577
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Simon Glass <sjg@google.com>