exymos: consolidate board_get_revision() for different platforms
There is no need to define this function in multiple places, it can be
kept in the common board file. As an enhancement, instead of using a
board specific size of the array of GPIOs, use the maximum number, as
the device tree contents will determine the actual number of pins to
probe.

BUG=chrome-os-partner:18329
TEST=manual

  . code was built and tested on the following platforms, reporting
    revisions as follows:
  snow             0
  peach-adv       39
  peach-rev1 4GB  31
  peach-rev1 2GB  76

Change-Id: I1a3466b013251377b932d7e33bd1c7c49915842f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/37689
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Simon Glass <sjg@google.com>
Commit-Queue: Vadim Bendebury <vbendeb@google.com>