exynos: Use more intuitive gpio ternary calculations
Exynos5 SOCs allow to control pull ups/pull downs internally, which in
turn, when proper value external resistors are used, allows to
determine if the GPIO pin configured for output has an external
pull up, pull down or not connected at all.

The gpio_decode_number() is doing it, but in a bit of a non-intuitive
way: logical one at the pin is mapped to value of two, logical zero is
mapped to one, and not connected state is mapped to zero.

A more intuitive way is use zero and one as is, and map 'not
connected' to two.

Another modification is to assigning different weights to the element
of list of GPIOs, the first element gets the highest wait - it is
natural when list elements are printed in one line (leftmost element
is the most significant).

The existing users of this function will be adjusted in the upcoming
patch.

BUG=chrome-os-partner:18329
TEST=manual
  . verified versions for different peach boards, they match the
    values assigned in the hardware description spreadsheet.

Change-Id: I9053f87979f3812e6e6940bb8e8fe137b299eb0c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/37731
Reviewed-by: Simon Glass <sjg@google.com>
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>