elog: Pass event log info to the kernel, and trim out the panic buffer
Pass the location of the event log panic buffer and the SPI flash
offset of the event log itself to the kernel through its device tree.

Also, trim the panic buffer range out of the /memory/reg property to
keep the kernel from using it for something else.

BRANCH=pit
BUG=chrome-os-partner:16132
TEST=Built and booted on pit. With prototyped panic code, verified
     that a panic event was stored in the right place, and was picked
     up by the firmware and recorded. Looked at the device tree in the
     booted system and verified that the expected properties where
     present in the right places. BRANCH=pit

Change-Id: I54e0dc2e26c14b2a86fb0ab460df243152dc50c7
Signed-off-by: Gabe Black <gabeblack@google.com>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/66782
Reviewed-by: Simon Glass <sjg@chromium.org>